+++ rs/rs_api_test.go
@@ -25,6 +25,10 @@ func init() {
 		panic("require ACCESS_KEY & SECRET_KEY")
 	}
 	client = New(nil)
+
+	// 删除 可能存在的 newkey1  newkey2 
+	client.Delete(nil, bucketName, newkey1)
+	client.Delete(nil, bucketName, newkey2)
 }
 
 func TestGetPrivateUrl(t *testing.T) {